Overhaul: simplify and improve code reuse
